+++ test/chart/bar_chart/bar_chart_data_test.dart
@@ -43,11 +43,11 @@ void main() {
           false);
       expect(
           barChartRodStackItem1 ==
-              barChartRodStackItem1Clone.copyWith(color: Colors.red),
+              barChartRodStackItem1Clone.copyWith(colors: [Colors.red]),
           false);
       expect(
           barChartRodStackItem1 ==
-              barChartRodStackItem1Clone.copyWith(color: Colors.green),
+              barChartRodStackItem1Clone.copyWith(colors: [Colors.green]),
           true);
       expect(barChartRodStackItem1 == barChartRodStackItem2, false);
     });