+++ test/chart/data_pool.dart
@@ -2542,7 +2542,7 @@ final ScatterChartData scatterChartData1Clone = ScatterChartData(
 final BarChartRodStackItem barChartRodStackItem1 = BarChartRodStackItem(
   1,
   2,
-  Colors.green,
+  [Colors.green],
 );
 final BarChartRodStackItem barChartRodStackItem1Clone =
     barChartRodStackItem1.copyWith();
@@ -2550,7 +2550,7 @@ final BarChartRodStackItem barChartRodStackItem1Clone =
 final BarChartRodStackItem barChartRodStackItem2 = BarChartRodStackItem(
   2,
   3,
-  Colors.green,
+  [Colors.green],
 );
 
 final BackgroundBarChartRodData backgroundBarChartRodData1 =